refactor: split the include machinery into one module per concern
The include family had grown into a single sixteen-hundred-line file mixing the
form-token spine, the commit strategies, three registration builders and the
per-form mounts. Each concern moves to its own module behind the same public
paths: the mount tokens and the commit trait stay crate-internal, and the
builders gain constructors instead of exposing their fields to sibling modules.
