+++ src/capability/tests.rs
@@ -1,4 +1,3 @@
-
 use super::*;
 
 #[test]
