+++ src/memory/capability/tests.rs
@@ -1,4 +1,3 @@
-
 use futures::StreamExt;
 
 use super::super::{MemoryBroker, MemorySource};
