+++ src/memory/tests.rs
@@ -1,4 +1,3 @@
-
 use futures::StreamExt;
 
 use super::*;
