+++ src/runtime/app/include/forms_eager.rs
@@ -1,0 +1,167 @@
+//! Eager mount forms: plain, raw, attachment-free injections and plain batch.
+
+// The typed default-reply commits need a default codec, so that import is gated the same way;
+// the raw default-reply commit publishes bare bytes and needs only `DefaultPublish`.
+// The default-reply commits build a `TypedPublisher` over the broker's default policy, so those
+// imports are gated with the default codec they require.
+use crate::{BatchSubscriber, Broker, BuildContext, Connected, Subscriber, SubscriptionSource};
+
+use crate::runtime::SliceHandler;
+use crate::runtime::batch::BatchDef;
+use crate::runtime::handler::Handler;
+use crate::runtime::inject::{FromStartup, InjectCall, InjectHandler};
+use crate::runtime::input::{DecodeWith, InputKind, RawBytes};
+use crate::runtime::middleware::Layer;
+use crate::runtime::subscriber_def::SubscriberDef;
+use crate::runtime::typed::Typed;
+
+use super::{IncludeMount, ScopeCodec, forms};
+use crate::runtime::app::scope::BrokerScope;
+
+// ---------------------------------------------------------------------------------------------
+// Plain subscribing: eager, no builder.
+
+impl<'s, B, Layers, C, State, Pipeline, Def> IncludeMount<'s, B, Layers, C, State, Pipeline, Def>
+    for forms::Subscribing
+where
+    B: Broker + 'static,
+    C: ScopeCodec,
+    Def: SubscriberDef,
+    Def::Source: SubscriptionSource<Connected<B>> + Send + 'static,
+    <Def::Source as SubscriptionSource<Connected<B>>>::Subscriber: Send + 'static,
+    <<Def::Source as SubscriptionSource<Connected<B>>>::Subscriber as Subscriber>::Message: 'static,
+    Def::Input: DecodeWith<<C as ScopeCodec>::Codec>,
+    Def::Handler: 'static,
+    Def::Context: BuildContext<
+            <<Def::Source as SubscriptionSource<Connected<B>>>::Subscriber as Subscriber>::Message,
+        > + Send
+        + 'static,
+    State: Send + Sync + 'static,
+    Layers: Layer<
+        Typed<
+            <<Def::Source as SubscriptionSource<Connected<B>>>::Subscriber as Subscriber>::Message,
+            Def::Input,
+            C::Codec,
+            Def::Handler,
+        >,
+    >,
+    Layers::Handler: Handler<
+            <<Def::Source as SubscriptionSource<Connected<B>>>::Subscriber as Subscriber>::Message,
+            Def::Context,
+            State,
+        > + 'static,
+{
+    type Out = ();
+
+    fn begin(def: Def, scope: &'s mut BrokerScope<B, Layers, C, State, Pipeline>) {
+        let source = def.source();
+        let codec = scope.codec.scope_codec();
+        scope.mount_subscriber(source, def, codec);
+    }
+}
+
+// ---------------------------------------------------------------------------------------------
+// Raw subscribing: eager, no builder, and no codec anywhere - the byte input kind decodes with
+// `()`, so the scope codec parameter `C` is left unconstrained and the mount works without any
+// codec feature.
+
+impl<'s, B, Layers, C, State, Pipeline, Def> IncludeMount<'s, B, Layers, C, State, Pipeline, Def>
+    for forms::RawSubscribing
+where
+    B: Broker + 'static,
+    Def: SubscriberDef<Input = RawBytes>,
+    Def::Source: SubscriptionSource<Connected<B>> + Send + 'static,
+    <Def::Source as SubscriptionSource<Connected<B>>>::Subscriber: Send + 'static,
+    <<Def::Source as SubscriptionSource<Connected<B>>>::Subscriber as Subscriber>::Message: 'static,
+    Def::Handler: 'static,
+    Def::Context: BuildContext<
+            <<Def::Source as SubscriptionSource<Connected<B>>>::Subscriber as Subscriber>::Message,
+        > + Send
+        + 'static,
+    State: Send + Sync + 'static,
+    Layers: Layer<
+        Typed<
+            <<Def::Source as SubscriptionSource<Connected<B>>>::Subscriber as Subscriber>::Message,
+            RawBytes,
+            (),
+            Def::Handler,
+        >,
+    >,
+    Layers::Handler: Handler<
+            <<Def::Source as SubscriptionSource<Connected<B>>>::Subscriber as Subscriber>::Message,
+            Def::Context,
+            State,
+        > + 'static,
+{
+    type Out = ();
+
+    fn begin(def: Def, scope: &'s mut BrokerScope<B, Layers, C, State, Pipeline>) {
+        let source = def.source();
+        scope.mount_subscriber(source, def, ());
+    }
+}
+
+// ---------------------------------------------------------------------------------------------
+// Attachment-free injections: eager, no builder - a definition whose startup injections need
+// nothing from the include site (a Seek parameter without an Out one) resolves against the
+// subscription itself.
+
+impl<'s, B, Layers, C, State, Pipeline, Def> IncludeMount<'s, B, Layers, C, State, Pipeline, Def>
+    for forms::Seek
+where
+    B: Broker + 'static,
+    C: ScopeCodec,
+    Def: InjectCall<State> + 'static,
+    Def::Source: SubscriptionSource<Connected<B>> + Send + 'static,
+    <Def::Source as SubscriptionSource<Connected<B>>>::Subscriber: Sync + Send + 'static,
+    <<Def::Source as SubscriptionSource<Connected<B>>>::Subscriber as Subscriber>::Message:
+        Send + Sync + 'static,
+    Def::Input: DecodeWith<<C as ScopeCodec>::Codec>,
+    Def::Context: BuildContext<
+            <<Def::Source as SubscriptionSource<Connected<B>>>::Subscriber as Subscriber>::Message,
+        > + Send
+        + Sync
+        + 'static,
+    Def::Injections: FromStartup<B, <Def::Source as SubscriptionSource<Connected<B>>>::Subscriber, ((),)>
+        + Send
+        + Sync
+        + 'static,
+    State: Send + Sync + 'static,
+    Layers: Layer<InjectHandler<Def, <C as ScopeCodec>::Codec>> + Clone + Send + 'static,
+    Layers::Handler: Handler<
+            <<Def::Source as SubscriptionSource<Connected<B>>>::Subscriber as Subscriber>::Message,
+            Def::Context,
+            State,
+        > + 'static,
+{
+    type Out = ();
+
+    fn begin(def: Def, scope: &'s mut BrokerScope<B, Layers, C, State, Pipeline>) {
+        let source = def.source();
+        scope.mount_inject(source, def, ((),));
+    }
+}
+
+// ---------------------------------------------------------------------------------------------
+// Plain batch: eager, no builder.
+
+impl<'s, B, Layers, C, State, Pipeline, Def> IncludeMount<'s, B, Layers, C, State, Pipeline, Def>
+    for forms::Batch
+where
+    B: Broker + 'static,
+    C: ScopeCodec,
+    Def: BatchDef,
+    Def::Source: SubscriptionSource<Connected<B>> + Send + 'static,
+    <Def::Source as SubscriptionSource<Connected<B>>>::Subscriber: BatchSubscriber + Send + 'static,
+    Def::Input: DecodeWith<<C as ScopeCodec>::Codec>,
+    Def::Handler: SliceHandler<<Def::Input as InputKind>::Owned, State> + 'static,
+    State: Send + Sync + 'static,
+{
+    type Out = ();
+
+    fn begin(def: Def, scope: &'s mut BrokerScope<B, Layers, C, State, Pipeline>) {
+        let source = def.source();
+        let codec = scope.codec.scope_codec();
+        scope.mount_batch(source, def, codec);
+    }
+}
