+++ src/runtime/batch_publishing/tests.rs
@@ -1,4 +1,3 @@
-
 use std::sync::Arc;
 use std::sync::atomic::{AtomicUsize, Ordering};
