+++ src/runtime/slot/tests.rs
@@ -1,4 +1,3 @@
-
 use std::collections::HashMap;
 
 use super::*;
