+++ src/typed_headers/ser/tests.rs
@@ -1,4 +1,3 @@
-
 use std::collections::BTreeMap;
 
 use serde::Serialize;
